move guard normalizer to own file
